dev-requirements.txt
isort==5.10.1
pylint==2.14.0
pylint-django==2.5.3
black==22.3.0
pdbpp==0.10.3
pytest-django==4.5.2
ipython==8.4.0